requirements.txt
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
fake-useragent==0.1.11
html5lib==1.0.1
idna==2.8
lxml==4.4.2
mysql-connector==2.2.9
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
urllib3==1.25.7
webencodings==0.5.1
